# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"AdminResLogin": {
				"properties": {
					"token_key": {
						"type": "string"
					},
					"message": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"AdminReqLogin": {
				"properties": {
					"name": {
						"type": "string",
						"enum": [
							"admin"
						],
						"nullable": false
					},
					"email": {
						"type": "string",
						"enum": [
							"admin@test.com"
						],
						"nullable": false
					},
					"password": {
						"type": "string",
						"enum": [
							"admin@123"
						],
						"nullable": false
					}
				},
				"required": [
					"name",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveResUser": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"pictureUrl": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"username",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveReqUser": {
				"properties": {
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"pictureUrl": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DeleteReqUser": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"_LeanDocument__LeanDocument_T__": {
				"properties": {},
				"type": "object"
			},
			"Pick__LeanDocument_T_.Exclude_keyof_LeanDocument_T_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested__": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/_LeanDocument__LeanDocument_T__",
						"description": "This documents _id."
					},
					"__v": {
						"description": "This documents __v."
					},
					"id": {
						"description": "The string version of this documents _id."
					}
				},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit__LeanDocument_this_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested_": {
				"$ref": "#/components/schemas/Pick__LeanDocument_T_.Exclude_keyof_LeanDocument_T_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"LeanDocument_this_": {
				"$ref": "#/components/schemas/Omit__LeanDocument_this_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested_"
			},
			"IUser": {
				"$ref": "#/components/schemas/LeanDocument_this_"
			},
			"UserReqLogin": {
				"properties": {
					"userId": {
						"type": "string"
					}
				},
				"required": [
					"userId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateReqUser": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"pictureUrl": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"username",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IMsgs": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"msg": {
						"type": "string"
					}
				},
				"required": [
					"userId",
					"msg"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveResGroup": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"users": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"messages": {
						"items": {
							"$ref": "#/components/schemas/IMsgs"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"name",
					"users",
					"messages"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveReqGroup": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DeleteReqGroup": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AddUserReqGroup": {
				"properties": {
					"groupId": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					}
				},
				"required": [
					"groupId",
					"userId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CheckMsgReqGroup": {
				"properties": {
					"messages": {
						"items": {
							"$ref": "#/components/schemas/IMsgs"
						},
						"type": "array"
					}
				},
				"required": [
					"messages"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveMessageRes": {
				"properties": {
					"groupId": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"groupId",
					"userId",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveMessageReq": {
				"properties": {
					"groupId": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"groupId",
					"userId",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "token",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "swagger-api-test",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/admin/saveAdmin": {
			"post": {
				"operationId": "SaveAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdminResLogin"
								}
							}
						}
					}
				},
				"tags": [
					"admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AdminReqLogin"
							}
						}
					}
				}
			}
		},
		"/user/saveUser": {
			"post": {
				"operationId": "SaveUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveResUser"
								}
							}
						}
					}
				},
				"tags": [
					"user"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveReqUser"
							}
						}
					}
				}
			}
		},
		"/user/deleteUser": {
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "User deleted",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveResUser"
								}
							}
						}
					}
				},
				"tags": [
					"user"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteReqUser"
							}
						}
					}
				}
			}
		},
		"/user/userLogin": {
			"post": {
				"operationId": "UserLogin",
				"responses": {
					"200": {
						"description": "User login",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IUser"
										},
										{
											"properties": {
												"_id": {
													"type": "string"
												}
											},
											"required": [
												"_id"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"user"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserReqLogin"
							}
						}
					}
				}
			}
		},
		"/user/updateUser": {
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveResUser"
								}
							}
						}
					}
				},
				"tags": [
					"user"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateReqUser"
							}
						}
					}
				}
			}
		},
		"/group/saveGroup": {
			"post": {
				"operationId": "SaveGroup",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveResGroup"
								}
							}
						}
					}
				},
				"tags": [
					"group"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveReqGroup"
							}
						}
					}
				}
			}
		},
		"/group/deleteGroup": {
			"delete": {
				"operationId": "DeleteGroup",
				"responses": {
					"200": {
						"description": "Group deleted",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveResGroup"
								}
							}
						}
					}
				},
				"tags": [
					"group"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteReqGroup"
							}
						}
					}
				}
			}
		},
		"/group/addUsers": {
			"post": {
				"operationId": "AddUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveResGroup"
								}
							}
						}
					}
				},
				"tags": [
					"group"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddUserReqGroup"
							}
						}
					}
				}
			}
		},
		"/group/checkMsg": {
			"post": {
				"operationId": "CheckMsg",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveResGroup"
								}
							}
						}
					}
				},
				"tags": [
					"group"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CheckMsgReqGroup"
							}
						}
					}
				}
			}
		},
		"/group/sendMessage": {
			"post": {
				"operationId": "SendMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveMessageRes"
								}
							}
						}
					}
				},
				"tags": [
					"group"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveMessageReq"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}